spritz up membership requests list
