+++ app/models/concerns/avatar_initials.rb
@@ -8,11 +8,12 @@ module AvatarInitials
   extend ActiveSupport::Concern
 
   def set_avatar_initials
-    initials = ""
     if name.blank? || name == email
       initials = email[0..1]
     else
-      name.split.each { |name| initials += name[0] }
+      initials = name.split.
+                      map {|name| name[0] }.
+                      join('')
     end
     initials = initials.upcase.gsub(/ /, '')
     initials = "DU" if deleted_at
