+++ app/views/groups/manage_membership_requests/_requestor.html.haml
@@ -1,0 +1,10 @@
+-if request.requestor.present?
+  %td= render 'base/linked_avatar', user: request.requestor
+  %td
+    =link_to request.name, user_path(request.requestor)
+    =request.email
+-else
+  %td= render 'base/avatar', user: LoggedOutUser.new(name: request.name, email: request.email)
+  %td
+    #{request.name} |
+    #{request.email}
