+++ app/views/groups/manage_membership_requests/index.html.haml
@@ -1,21 +1,36 @@
 - set_title @group.name, "", @group.parent
 = render '/groups/title', group: @group
-%h2 Membership requests
-%br
--if can? :add_members, @group
-  -pending_requests = @group.membership_requests.where('response IS ?', nil)
-  -if pending_requests.exists?
-    %table.table#membership-request-list
+
+%h2=t :membership_requests
+-if @current_requests.present?
+  %table.table#membership-request-list
+    -@current_requests.each do |request|
       %tr
-        %th #{t(:name)} & #{t('common.email')}
-        %th Introduction
-        %th= " "#t(:approve)
-        %th= " "#t(:ignore_request)
-      -pending_requests.where('response IS ?', nil).each do |request|
-        %tr
-          %td #{request.name} &lt;#{request.email}&gt;
-          %td= render_rich_text(request.introduction, false)
-          %td= link_to t(:approve), approve_membership_request_path(request), method: :post, id: "approve-membership-request-#{request.id}"
-          %td= link_to t(:ignore), ignore_membership_request_path(request), method: :post, id: "ignore-membership-request-#{request.id}"
-  -else
-    = t(:no_pending_membership_requests)
+        =render 'requestor', request: request
+        %td
+          =render_rich_text(request.introduction, false)
+          %span.time-ago-subtle= t(:ago, quantity_of_time: time_ago_in_words(request.created_at))
+
+        %td
+          =link_to approve_membership_request_path(request), method: :post, id: "approve-membership-request-#{request.id}" do
+            %button.btn=t(:approve)
+          =t(:or) + ' '
+          =link_to t(:ignore).downcase, ignore_membership_request_path(request), method: :post, id: "ignore-membership-request-#{request.id}"
+-else
+  .no-current-requests= t(:no_pending_membership_requests)
+
+%h2=t :previous_requests
+%table.table
+  - @previous_requests.each do |request|
+    %tr
+      =render 'requestor', request: request
+      %td= render_rich_text(request.introduction, false)
+      %td.span2
+        -responder = request.responder.name
+        -case request.response
+        -when 'approved'
+          =t(:approved_by_person, person: responder)
+        -when 'ignored'
+          =t(:ignored_by_person, person: responder)
+        %span.time-ago-subtle= t(:ago, quantity_of_time: time_ago_in_words(request.responded_at))
+= paginate @previous_requests
